results/rebuttal_ablation_no_summarization/run-20231115-165815_LLMatic_donnemartin-system-design-oop-whatsapp_0_20-runs_log/whatsapp/7/LLMatic/app.py
from flask import Flask, request
import hashlib

app = Flask(__name__)

# Mock database
DATABASE = {
	'users': {},
	'messages': {},
	'groups': {},
	'statuses': {},
	'profile_pictures': {},
	'privacy_settings': {},
	'blocked_contacts': {},
	'read_receipts': {},
	'images': {}
}

@app.route('/')

def home():
	return 'Hello, World!'

@app.route('/register', methods=['POST'])

def register():
	email = request.json.get('email')
	password = request.json.get('password')
	if email in DATABASE['users']:
		return {'message': 'User already exists'}, 400
	DATABASE['users'][email] = {'password': password}
	return {'message': 'User registered successfully'}, 200

@app.route('/forgot_password', methods=['POST'])

def forgot_password():
	email = request.json.get('email')
	if email not in DATABASE['users']:
		return {'message': 'User does not exist'}, 400
	# In a real application, we would generate a secure link and email it to the user.
	# For this mock application, we'll just return a dummy link.
	return {'recovery_link': 'http://example.com/recover?email=' + email}, 200

@app.route('/set_profile_picture', methods=['POST'])

def set_profile_picture():
	user_id = request.json.get('user_id')
	new_picture = request.json.get('new_picture')
	DATABASE['profile_pictures'][user_id] = new_picture
	return {'message': 'Profile picture updated successfully'}, 200

@app.route('/set_status_message', methods=['POST'])

def set_status_message():
	user_id = request.json.get('user_id')
	new_message = request.json.get('new_message')
	DATABASE['statuses'][user_id] = new_message
	return {'message': 'Status message updated successfully'}, 200

@app.route('/update_privacy_settings', methods=['POST'])

def update_privacy_settings():
	user_id = request.json.get('user_id')
	new_settings = request.json.get('new_settings')
	DATABASE['privacy_settings'][user_id] = new_settings
	return {'message': 'Privacy settings updated successfully'}, 200

@app.route('/block_contact', methods=['POST'])

def block_contact():
	user_id = request.json.get('user_id')
	contact_id = request.json.get('contact_id')
	DATABASE['blocked_contacts'].setdefault(user_id, []).append(contact_id)
	return {'message': 'Contact blocked successfully'}, 200

@app.route('/unblock_contact', methods=['POST'])

def unblock_contact():
	user_id = request.json.get('user_id')
	contact_id = request.json.get('contact_id')
	if contact_id in DATABASE['blocked_contacts'].get(user_id, []):
		DATABASE['blocked_contacts'][user_id].remove(contact_id)
	return {'message': 'Contact unblocked successfully'}, 200

@app.route('/create_group', methods=['POST'])

def create_group():
	group_id = request.json.get('group_id')
	group_data = request.json.get('group_data')
	group_data['participants'] = []
	DATABASE['groups'][group_id] = group_data
	return {'message': 'Group created successfully'}, 200

@app.route('/edit_group', methods=['POST'])

def edit_group():
	group_id = request.json.get('group_id')
	group_data = request.json.get('group_data')
	if group_id not in DATABASE['groups']:
		return {'message': 'Group does not exist'}, 400
	DATABASE['groups'][group_id] = group_data
	return {'message': 'Group edited successfully'}, 200

@app.route('/add_participant', methods=['POST'])

def add_participant():
	group_id = request.json.get('group_id')
	participant_id = request.json.get('participant_id')
	if group_id not in DATABASE['groups']:
		return {'message': 'Group does not exist'}, 400
	if 'participants' not in DATABASE['groups'][group_id]:
		return {'message': 'Group data is missing participants key'}, 400
	DATABASE['groups'][group_id]['participants'].append(participant_id)
	return {'message': 'Participant added successfully'}, 200

@app.route('/remove_participant', methods=['POST'])

def remove_participant():
	group_id = request.json.get('group_id')
	participant_id = request.json.get('participant_id')
	if group_id not in DATABASE['groups']:
		return {'message': 'Group does not exist'}, 400
	if 'participants' not in DATABASE['groups'][group_id]:
		return {'message': 'Group data is missing participants key'}, 400
	if participant_id in DATABASE['groups'][group_id]['participants']:
		DATABASE['groups'][group_id]['participants'].remove(participant_id)
	return {'message': 'Participant removed successfully'}, 200

@app.route('/manage_admin', methods=['POST'])

def manage_admin():
	group_id = request.json.get('group_id')
	admin_data = request.json.get('admin_data')
	if group_id not in DATABASE['groups']:
		return {'message': 'Group does not exist'}, 400
	DATABASE['groups'][group_id]['admin'] = admin_data
	return {'message': 'Admin roles and permissions updated successfully'}, 200

@app.route('/send_message', methods=['POST'])

def send_message():
	message_id = request.json.get('message_id')
	message = request.json.get('message')
	DATABASE['messages'][message_id] = message
	return {'message': 'Message sent successfully'}, 200

@app.route('/read_receipt', methods=['POST'])

def read_receipt():
	message_id = request.json.get('message_id')
	DATABASE['read_receipts'][message_id] = True
	return {'message': 'Read receipt updated successfully'}, 200

@app.route('/encrypt_message', methods=['POST'])

def encrypt_message():
	message = request.json.get('message')
	encrypted_message = hashlib.sha256(message.encode()).hexdigest()
	return {'encrypted_message': encrypted_message}, 200

@app.route('/share_image', methods=['POST'])

def share_image():
	image_id = request.json.get('image_id')
	image = request.json.get('image')
	DATABASE['images'][image_id] = image
	return {'message': 'Image shared successfully'}, 200

if __name__ == '__main__':
	app.run(debug=True)
